% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chattragent_use.R
\name{chattragent_use}
\alias{chattragent_use}
\title{chattragent_use}
\usage{
chattragent_use(agent)
}
\arguments{
\item{agent}{the agent name in the dataset}
}
\value{
agent
}
\description{
chattr agent setting
}
